requirements.txt
google-adk
alpha-vantage
python-dotenv==1.0.0
requests
